refactor StatusContext
