+++ src/components/Auth/PrivateRoute.tsx
@@ -6,7 +6,7 @@ import { useTranslation, Trans } from 'react-i18next';
 
 import Spinner from '../Shared/Spinner'; // Import your spinner component
 import { useSessionStorage } from '../../hooks/useStorage';
-import StatusContext from '../../context/StatusContext';
+import StatusContext from '@/context/StatusContext';
 import SessionContext from '@/context/SessionContext';
