+++ src/components/Credentials/CredentialDeleteButton.jsx
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 import { MdDelete } from 'react-icons/md';
 import { useTranslation } from 'react-i18next';
 import Button from '../Buttons/Button';
-import StatusContext from '../../context/StatusContext';
+import StatusContext from '@/context/StatusContext';
 
 const CredentialDeleteButton = ({ onDelete }) => {
 	const { t } = useTranslation();
