+++ src/components/Notifications/UpdateNotification.jsx
@@ -1,5 +1,5 @@
 import React, { useContext, useState } from 'react';
-import StatusContext from '../../context/StatusContext';
+import StatusContext from '@/context/StatusContext';
 import { useTranslation } from 'react-i18next';
 import Button from '../Buttons/Button';
 import { FaTimes } from 'react-icons/fa';
