+++ src/lib/services/OpenID4VCIHelper.tsx
@@ -6,7 +6,7 @@ import { base64url, importX509, jwtVerify } from "jose";
 import { getPublicKeyFromB64Cert } from "../utils/pki";
 import { useHttpProxy } from "./HttpProxy/HttpProxy";
 import { useCallback, useContext, useEffect, useState, useMemo } from "react";
-import StatusContext from "../../context/StatusContext";
+import StatusContext from "@/context/StatusContext";
 import SessionContext from "@/context/SessionContext";
 
 export function useOpenID4VCIHelper(): IOpenID4VCIHelper {
