+++ src/pages/Login/Login.tsx
@@ -8,7 +8,7 @@ import { Trans, useTranslation } from 'react-i18next';
 import type { CachedUser } from '../../services/LocalStorageKeystore';
 import { calculateByteSize } from '../../util';
 
-import StatusContext from '../../context/StatusContext';
+import StatusContext from '@/context/StatusContext';
 import SessionContext from '@/context/SessionContext';
 
 import * as config from '../../config';
