+++ src/pages/Login/LoginState.jsx
@@ -4,7 +4,7 @@ import { FaInfoCircle } from 'react-icons/fa';
 import { GoPasskeyFill } from 'react-icons/go';
 import { Trans, useTranslation } from 'react-i18next';
 
-import StatusContext from '../../context/StatusContext';
+import StatusContext from '@/context/StatusContext';
 import SessionContext from '@/context/SessionContext';
 
 import LanguageSelector from '../../components/LanguageSelector/LanguageSelector';
