+++ src/pages/Settings/Settings.tsx
@@ -5,7 +5,7 @@ import { BsLock, BsPlusCircle, BsUnlock } from 'react-icons/bs';
 import { MdNotifications } from "react-icons/md";
 import { IoIosArrowDown } from "react-icons/io";
 
-import StatusContext from '../../context/StatusContext';
+import StatusContext from '@/context/StatusContext';
 import SessionContext from '@/context/SessionContext';
 
 import useScreenType from '../../hooks/useScreenType';
